requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-google-genai
langchain-community
langgraph
tavily-python
pydantic
python-dotenv
python-multipart